+++ wand/cdefs/magick_image.py
@@ -1210,6 +1210,14 @@ def load(lib, IM_VERSION):
             lib.MagickWaveletDenoiseImage = None
     else:
         lib.MagickWaveletDenoiseImage = None
+    if IM_VERSION > 0x709:
+        try:
+            lib.MagickWhiteBalanceImage.argtypes = [c_void_p]
+            lib.MagickWhiteBalanceImage.restype = c_bool
+        except AttributeError:
+            lib.MagickWhiteBalanceImage = None
+    else:
+        lib.MagickWhiteBalanceImage = None
     lib.MagickWhiteThresholdImage.argtypes = [c_void_p, c_void_p]
     lib.MagickWhiteThresholdImage.restype = c_bool
     lib.MagickWriteImage.argtypes = [c_void_p, c_char_p]